Add ability to apply DataIO to Data after construction (#159)
* move region slicing to specific module

* add ability to wrap Data after construction

* add test

* satisfy flake8

* move tests

* add edge case test

* satisfy flake8
