+++ src/hdmf/__init__.py
@@ -1,7 +1,7 @@
 from . import query  # noqa: F401
 from .container import Container, Data, DataRegion
 from .utils import docval, getargs
-from .data_utils import ListSlicer
+from .region import ListSlicer
 from .backends.hdf5.h5_utils import H5RegionSlicer, H5Dataset
