+++ src/hdmf/data_utils.py
@@ -4,14 +4,11 @@ try:
 except ImportError:
     from collections import Iterable  # Python 2.7
 
-from operator import itemgetter
-
 import numpy as np
 from warnings import warn
 from six import with_metaclass, text_type, binary_type
 import copy
 
-from .container import Data, DataRegion
 from .utils import docval, getargs, popargs, docval_macro, get_data_shape
 
 
@@ -597,7 +594,7 @@ class DataIO(with_metaclass(ABCMeta, object)):
     Base class for wrapping data arrays for I/O. Derived classes of DataIO are typically
     used to pass dataset-specific I/O parameters to the particular HDMFIO backend.
     """
-    @docval({'name': 'data', 'type': 'array_data', 'doc': 'the data to be written'})
+    @docval({'name': 'data', 'type': 'array_data', 'doc': 'the data to be written', 'default': None})
     def __init__(self, **kwargs):
         data = popargs('data', kwargs)
         self.__data = data
@@ -606,6 +603,13 @@ class DataIO(with_metaclass(ABCMeta, object)):
     def data(self):
         """Get the wrapped data object"""
         return self.__data
+
+    @data.setter
+    def data(self, val):
+        """Set the wrapped data object"""
+        if self.__data is not None:
+            raise ValueError("cannot overwrite 'data' on DataIO")
+        self.__data = val
 
     def __copy__(self):
         """
@@ -688,91 +692,6 @@ class DataIO(with_metaclass(ABCMeta, object)):
     def valid(self):
         """bool indicating if the data object is valid"""
         return self.data is not None
-
-
-class RegionSlicer(with_metaclass(ABCMeta, DataRegion)):
-    '''
-    A abstract base class to control getting using a region
-
-    Subclasses must implement `__getitem__` and `__len__`
-    '''
-
-    @docval({'name': 'target', 'type': None, 'doc': 'the target to slice'},
-            {'name': 'slice', 'type': None, 'doc': 'the region to slice'})
-    def __init__(self, **kwargs):
-        self.__target = getargs('target', kwargs)
-        self.__slice = getargs('slice', kwargs)
-
-    @property
-    def data(self):
-        """The target data. Same as self.target"""
-        return self.target
-
-    @property
-    def region(self):
-        """The selected region. Same as self.slice"""
-        return self.slice
-
-    @property
-    def target(self):
-        """The target data"""
-        return self.__target
-
-    @property
-    def slice(self):
-        """The selected slice"""
-        return self.__slice
-
-    @property
-    @abstractmethod
-    def __getitem__(self, idx):
-        """Must be implemented by subclasses"""
-        pass
-
-    @property
-    @abstractmethod
-    def __len__(self):
-        """Must be implemented by subclasses"""
-        pass
-
-
-class ListSlicer(RegionSlicer):
-    """Implementation of RegionSlicer for slicing Lists and Data"""
-    @docval({'name': 'dataset', 'type': (list, tuple, Data), 'doc': 'the dataset to slice'},
-            {'name': 'region', 'type': (list, tuple, slice), 'doc': 'the region reference to use to slice'})
-    def __init__(self, **kwargs):
-        self.__dataset, self.__region = getargs('dataset', 'region', kwargs)
-        super(ListSlicer, self).__init__(self.__dataset, self.__region)
-        if isinstance(self.__region, slice):
-            self.__getter = itemgetter(self.__region)
-            self.__len = len(range(*self.__region.indices(len(self.__dataset))))
-        else:
-            self.__getter = itemgetter(*self.__region)
-            self.__len = len(self.__region)
-
-    def __read_region(self):
-        """
-        Internal helper function used to define self._read
-        """
-        if not hasattr(self, '_read'):
-            self._read = self.__getter(self.__dataset)
-            del self.__getter
-
-    def __getitem__(self, idx):
-        """
-        Get data values from selected data
-        """
-        self.__read_region()
-        getter = None
-        if isinstance(idx, (list, tuple)):
-            getter = itemgetter(*idx)
-        else:
-            getter = itemgetter(idx)
-        return getter(self._read)
-
-    def __len__(self):
-        """Number of values in the slice/region"""
-        return self.__len
 
 
 class InvalidDataIOError(Exception):
